changes, workflow läuft, tests fehlen

+++ src/com/company/store/ProcessingData.java
@@ -1,8 +1,8 @@
 package com.company.store;
 
 public class ProcessingData {
-    public static Long TOTAL_PAGES = 0L;
-    public static Long CURRENT_PAGE = 0L;
+    public static Integer TOTAL_PAGES = 0;
+    public static Integer CURRENT_PAGE = 0;
     public static String[] RAWLINES;
     public static String[] HEADLINE;
 }
